add DBTypes
git-svn-id: http://localhost:8080/svn/AlgoQuest/trunk@43 de6f410c-f7d5-3442-bb8a-973c24c52ac3

+++ AQUtil/aq/Utilities.cpp
@@ -159,7 +159,7 @@ int GetFiles( const char* pszSrcPath, std::vector<std::string>& files )
 //------------------------------------------------------------------------------
 char* ReadValidLine( FILE* pFIn, char* pszTmpBuf, int nSize, int nTrimEnd ) {
 	char *psz = NULL;
-	int  nLen;
+	size_t  nLen;
 
 	/* Read a valid line */
 	pszTmpBuf[ 0 ] = '\0';
